+++ src/main/java/de/uniks/beastopia/teaml/service/PresetsService.java
@@ -41,11 +41,11 @@ public class PresetsService {
                 .map((ResponseBody body) -> new Gson().fromJson(body.string(), TileSet.class));
     }
 
-    public Rectangle2D getTileViewPort(int id, TileSet tileSet) {
+    public Rectangle2D getTileViewPort(long id, TileSet tileSet) {
         int columns = tileSet.columns();
         int tileSize = tileSet.tileheight();
-        int x = (id - 1) % columns;
-        int y = (id - 1) / columns;
+        long x = (id - 1) % columns;
+        long y = (id - 1) / columns;
         return new Rectangle2D(x * tileSize, y * tileSize, tileSize, tileSize);
     }
